+++ packages/tacky-store/src/utils/decorator.ts
@@ -4,7 +4,7 @@ export function quacksLikeADecorator(args: any[]): boolean {
   return (args.length === 2 || args.length === 3) && typeof args[1] === 'string'
 }
 
-// prop can be decorated by @state
+// prop could be decorated by @reactor
 export const canObserve = (value: any): boolean => {
   return stateDecoRegExp.test(toObjectTypeString(value));
 }